Fixed clang 21 warnings about default initialization of an object of type

+++ kernel/bpf/syscall.c
@@ -1983,7 +1983,7 @@ static int bpf_prog_get_info_by_fd(struct file *file,
 				   const union bpf_attr *attr,
 				   union bpf_attr __user *uattr)
 {
-	struct bpf_prog_info __user *uinfo = u64_to_user_ptr(attr->info.info);
+	struct bpf_prog_info __user *uinfo = u64_to_user_ptr((u64)attr->info.info);
 	struct bpf_prog_info info;
 	u32 info_len = attr->info.info_len;
 	char __user *uinsns;
@@ -2163,7 +2163,7 @@ static int bpf_map_get_info_by_fd(struct file *file,
 				  const union bpf_attr *attr,
 				  union bpf_attr __user *uattr)
 {
-	struct bpf_map_info __user *uinfo = u64_to_user_ptr(attr->info.info);
+	struct bpf_map_info __user *uinfo = u64_to_user_ptr((u64)attr->info.info);
 	struct bpf_map_info info;
 	u32 info_len = attr->info.info_len;
 	int err;
@@ -2206,7 +2206,7 @@ static int bpf_btf_get_info_by_fd(struct file *file,
 				  const union bpf_attr *attr,
 				  union bpf_attr __user *uattr)
 {
-	struct bpf_btf_info __user *uinfo = u64_to_user_ptr(attr->info.info);
+	struct bpf_btf_info __user *uinfo = u64_to_user_ptr((u64)attr->info.info);
 	u32 info_len = attr->info.info_len;
 	int err;
 
@@ -2280,7 +2280,7 @@ static int bpf_task_fd_query_copy(const union bpf_attr *attr,
 				    const char *buf, u64 probe_offset,
 				    u64 probe_addr)
 {
-	char __user *ubuf = u64_to_user_ptr(attr->task_fd_query.buf);
+	char __user *ubuf = u64_to_user_ptr((u64)attr->task_fd_query.buf);
 	u32 len = buf ? strlen(buf) : 0, input_len;
 	int err = 0;
